pyproject.toml
[tool.poetry]
name = "simple-html"
version = "3.0.0"
readme = "README.md"
description = "Template-less html rendering in Python"
authors = ["Keith Philpott <fakekeith@example.org>"]
license = "MIT"
homepage = "https://github.com/keithasaurus/simple_html"
keywords = ["html", "type hints"]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Console',
    'Environment :: MacOS X',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'Operating System :: MacOS',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: POSIX :: Linux',
    'Operating System :: Unix',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Utilities',
    'Typing :: Typed'
]


[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
jinja2 = "3.1.6"
mypy = "1.17.1"
pytest = "8.4.1"
setuptools = "80.9.0"
fast-html = "1.0.12"
dominate = "2.9.1"
ruff = "0.12.8"


[build-system]
requires = [
    "poetry-core>=1.0.0",
    "mypy[mypyc]==1.17.1",
]

[tool.mypy]
allow_redefinition = false
allow_untyped_globals = false
check_untyped_defs = true
disallow_any_expr = false # currently need this for generics
disallow_any_generics = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
local_partial_types = true
no_implicit_optional = true
no_implicit_reexport = true
strict_optional = true
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_unreachable = true

[tool.cibuildwheel]
skip = ["cp38-*", "cp314*"]

[tool.cibuildwheel.windows]
archs = ["AMD64", "x86"]
